Restructure code better
