+++ app/javascript/src/contexts/auth.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import authReducer from "reducers/auth-reducer";
+import authReducer from "reducers/auth";
 
 const AuthStateContext = React.createContext();
 const AuthDispatchContext = React.createContext();
