+++ app/javascript/src/reducers/auth.js
