+++ app/views/home/index.html.slim
@@ -1,1 +1,1 @@
-= react_component("index", get_client_props, { class: 'root-container'})
+= react_component("App", get_client_props, { class: 'root-container'})
